transcribe_audio_function/requirements.txt
google-cloud-storage 
google-cloud-speech 